requirement.txt
Flask==2.3.0
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
Werkzeug==2.3.0
mysql-connector-python==9.1.0
sentence-transformers==3.3.1
qdrant-client==1.12.1
openai==1.55.1
pandas==2.2.3